websockets/interfaces/WebSocket/readyState/001.html

<!doctype html>
<title>WebSockets: getting readyState in connecting</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(t) {
  assert_equals((new WebSocket(SCHEME_AND_DOMAIN+'/')).readyState, WebSocket.CONNECTING);
});
</script>
